Lots of changes
Renamed the notebooks for clarity. Changed zmax determination to automatically get zmax, inf and sup, and saved this data to be used across the notebooks. Now zmax determination is separate from zmax det plotting. Created function to get the table and the parameters from a list of maximum redshift. Have to save it and get the mean values.